# backend/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "医院在线预约挂号系统 API",
    "description": "根据当前后端控制器、请求 DTO、测试断言与《项目分工与接口映射.md》整理的可导入 Apifox 的 OpenAPI 3.0 JSON。统一响应格式为 { code, message, data }。\n\n已补充 x-apifox-folder 和 operationId，导入后更容易按模块分组并生成接口结构。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发"
    }
  ],
...
  "paths": {
...
    "/api/articles": {
      "get": {
        "tags": [
          "文章"
        ],
        "summary": "文章列表",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageArticleResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "文章",
        "operationId": "getApiArticles",
        "x-apifox-summary": "文章列表"
      }
    },
    "/api/articles/{id}": {
      "get": {
        "tags": [
          "文章"
        ],
        "summary": "文章详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDetailResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "文章",
        "operationId": "getApiArticlesId",
        "x-apifox-summary": "文章详情"
      }
    },
...
    "/api/appointments": {
      "post": {
        "tags": [
          "挂号订单"
        ],
        "summary": "创建挂号订单",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentRequest"
              },
              "example": {
                "scheduleId": 1,
                "familyMemberId": 1,
                "diseaseDesc": "头晕"
              },
              "examples": {
                "createAppointment": {
                  "summary": "创建挂号订单",
                  "value": {
                    "scheduleId": 1,
                    "familyMemberId": 1,
                    "diseaseDesc": "头晕"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "orderNo": {
                          "type": "string",
                          "example": "AP202607170002"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "createAppointment": {
                    "summary": "创建挂号成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "orderNo": "AP202607170002"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "x-apifox-folder": "挂号订单",
        "operationId": "postApiAppointments",
        "x-apifox-summary": "创建挂号订单"
      }
    },
    "/api/appointments/my": {
      "get": {
        "tags": [
          "挂号订单"
        ],
        "summary": "我的挂号列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageAppointmentResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "挂号订单",
        "operationId": "getApiAppointmentsMy",
        "x-apifox-summary": "我的挂号列表"
      }
    },
    "/api/appointments/{orderNo}": {
      "get": {
        "tags": [
          "挂号订单"
        ],
        "summary": "挂号订单详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDetailResponse"
                },
                "examples": {
                  "appointmentDetail": {
                    "summary": "挂号订单详情",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "orderNo": "AP202607170001",
                        "doctor": {
                          "id": 1,
                          "name": "李主任"
                        },
                        "hospital": {
                          "id": 1,
                          "name": "北京第一人民医院"
                        },
                        "date": "2026-07-18",
                        "timeSlot": "08:00-09:00",
                        "patientName": "张三",
                        "status": 2,
                        "fee": 18.0,
                        "createTime": "2026-07-17T10:00:00"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "挂号订单",
        "operationId": "getApiAppointmentsOrderNo",
        "x-apifox-summary": "挂号订单详情"
      }
    },
    "/api/appointments/{orderNo}/cancel": {
      "post": {
        "tags": [
          "挂号订单"
        ],
        "summary": "取消挂号订单",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "examples": {
                  "successNull": {
                    "summary": "??",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "挂号订单",
        "operationId": "postApiAppointmentsOrderNoCancel",
        "x-apifox-summary": "取消挂号订单"
      }
    },
    "/api/appointments/{orderNo}/pay": {
      "post": {
        "tags": [
          "挂号订单"
        ],
        "summary": "挂号支付",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayRequest"
              },
              "example": {
                "payType": 2
              },
              "examples": {
                "payAppointment": {
                  "summary": "支付挂号订单",
                  "value": {
                    "payType": 2
                  }
                },
                "payAppointmentUnsupported": {
                  "summary": "不支持的支付方式",
                  "value": {
                    "payType": 9
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        },
                        "payUrl": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "payAppointmentSuccess": {
                    "summary": "支付成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "success": true,
                        "payUrl": null
                      }
                    }
                  },
                  "payAppointmentFail": {
                    "summary": "支付方式不支持",
                    "value": {
                      "code": 400,
                      "message": "支付方式不支持",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "挂号订单",
        "operationId": "postApiAppointmentsOrderNoPay",
        "x-apifox-summary": "挂号支付"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10
        }
      }
    },
...
    "schemas": {
...
      "CreateAppointmentRequest": {
        "type": "object",
        "required": [
          "scheduleId",
          "familyMemberId"
        ],
        "properties": {
          "scheduleId": {
            "type": "integer",
            "format": "int64"
          },
          "familyMemberId": {
            "type": "integer",
            "format": "int64"
          },
          "diseaseDesc": {
            "type": "string"
          }
        }
      },
...
      "PayRequest": {
        "type": "object",
        "required": [
          "payType"
        ],
        "properties": {
          "payType": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "PageArticleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "pageSize": {
                "type": "integer"
              }
            }
          }
        }
      },
      "ArticleDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "title": {
                "type": "string"
              },
              "summary": {
                "type": "string"
              },
              "content": {
                "type": "string"
              },
              "image": {
                "type": "string"
              },
              "author": {
                "type": "string"
              },
              "publishTime": {
                "type": "string"
              },
              "departmentName": {
                "type": "string"
              }
            }
          }
        }
      },
      "PageAppointmentResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "pageSize": {
                "type": "integer"
              }
            }
          }
        }
      },
      "AppointmentDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "orderNo": {
                "type": "string"
              },
              "doctor": {
                "type": "object"
              },
              "hospital": {
                "type": "object"
              },
              "date": {
                "type": "string",
                "format": "date"
              },
              "timeSlot": {
                "type": "string"
              },
              "patientName": {
                "type": "string"
              },
              "status": {
                "type": "integer"
              },
              "fee": {
                "type": "number"
              },
              "createTime": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}